# api/openapi.yaml
openapi: "3.0.3"

info:
  title: "API"
  version: "1.0.0"

servers:
  - url: http://{host}:{port}/
    description: development
    variables:
      host:
        default: localhost
      port:
        default: '8080'

paths:
  /ping:
    get:
      operationId: ping
      description: 'health check'
      tags:
        - healthCheck
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: pong
  /users:
    get:
      operationId: getUsers
      description: 'get users'
      tags:
        - user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
  /user/{userId}:
    parameters:
      - schema:
          type: string
          format: uuid
        name: userId
        in: path
        required: true
        description: 'user id'
    get:
      operationId: getUser
      description: 'get users'
      tags:
        - user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      operationId: updateUser
      description: 'update user'
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /user:
    post:
      operationId: createUser
      description: 'create user'
      tags:
        - user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
components:
  schemas:
    User:
      type: object
      required:
        - userId
        - userName
      properties:
        userId:
          format: uuid
          type: string
        userName:
          type: string
    Users:
      type: array
      items:
        $ref: '#/components/schemas/User'
    CreateUser:
      type: object
      required:
        - userName
      properties:
        userName:
          type: string